fix:kelolagenres_admin

+++ Development/laravel/database/migrations/2025_11_03_103255_create_films_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('films', function (Blueprint $table) {
-            $table->id();
-            $table->string('judul');
-            $table->text('sinopsis');
-            $table->integer('tahun_rilis');
-            $table->string('sutradara');
-            $table->string('aktor')->nullable();
-            $table->string('durasi')->nullable();
-            $table->string('poster')->nullable();
-            $table->foreignId('genre_id')->constrained('genres')->onDelete('cascade');
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('films')) {
+            Schema::create('films', function (Blueprint $table) {
+                $table->id();
+                $table->string('judul');
+                $table->text('sinopsis');
+                $table->integer('tahun_rilis');
+                $table->string('sutradara');
+                $table->string('aktor')->nullable();
+                $table->string('durasi')->nullable();
+                $table->string('poster')->nullable();
+                $table->foreignId('genre_id')->constrained('genres')->onDelete('cascade');
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
